pyproject.toml
[project]
name = "ai-engineering-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.37.1",
    "nltk>=3.9.2",
    "numpy>=2.3.5",
    "openai>=2.11.0",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.2.0",
    "whoosh>=2.7.4",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
